requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.6
openai==1.82.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
SQLAlchemy==2.0.41
